Fix flexbox and adjust styling
