+++ src/app.scss
@@ -2,9 +2,10 @@
 
 .header {
   height: 40px;
-  margin: 30px;
+  margin: 10px;
+  margin-bottom: 20px;
   font-family: 'Trade Winds', cursive;
-  text-align: center;
+
   display: flex;
   align-items: center;
   justify-content: center;
@@ -26,7 +27,9 @@
 .wrapper {
   display: flex;
   flex-flow: row nowrap;
-
-  align-items: flex-start;
-  justify-content: space-between;
+  justify-content: flex-start;
+  align-items: stretch;
+  overflow: hidden;
+  margin: 10px;
+  width: 100%;
 }
