+++ src/components/favourites.scss
@@ -1,29 +1,26 @@
+.favouriteListContainerWrapper {
+  background-color: rgba(186, 186, 231, 0.877);
+  border-radius: 30px;
+  margin-right: 10px;
+  border: 5px solid rgba(74, 133, 189, 0.534);
+  overflow: scroll;
+  max-width: 150px;
+  flex-shrink: 0;
+  flex-grow: 1;
+  flex-basis: auto;
+  height: 100%;
+}
+
 .favouriteListContainer {
   display: flex;
   flex-flow: row wrap;
-  align-items: space-around;
-
-  background-color: rgba(186, 186, 231, 0.658);
-  border-radius: 50px;
-
-  background-color: rgb(105, 105, 202);
-
-  margin-right: 10px;
-  padding: 10px;
-  max-height: 200px;
-  overflow: auto;
-
-  flex-grow: 1;
-  flex-shrink: 0;
-  overflow-y: auto;
-  background-color: rgba(186, 186, 231, 0.658);
-  border-radius: 50px;
-  margin-right: 10px;
-  padding: 10px;
+  justify-content: space-around;
+  width: 100%;
+  height: 100%;
 }
 
 .favouriteCard {
-  margin: 20px;
+  margin: 10px;
   text-align: center;
   border: 2px solid rgba(15, 15, 34, 0.61);
   padding: 10px;
