+++ src/components/searchresults.scss
@@ -1,0 +1,33 @@
+.listContainerWrapper {
+  flex-grow: 1;
+  text-align: center;
+  overflow: scroll;
+  margin: 10px;
+}
+
+.listContainer {
+  display: flex;
+  justify-content: space-around;
+  align-items: center;
+  flex-flow: row wrap;
+  width: 100%;
+  height: 100%;
+}
+
+.resultItem {
+  margin: 10px;
+  text-align: center;
+  border: 2px solid rgba(15, 15, 34, 0.61);
+  padding: 10px;
+  border-radius: 0.4rem;
+  background: rgb(74, 133, 189);
+
+  &:hover {
+    background-color: rgb(142, 163, 221);
+    cursor: pointer;
+  }
+
+  &:active {
+    background-color: rgb(72, 39, 192);
+  }
+}
