+++ src/components/title.scss
@@ -1,7 +1,7 @@
 .title {
   margin: 0;
-  font-size: 3.5rem;
-  letter-spacing: 0.3em;
+  font-size: 2.5rem;
+  letter-spacing: 0.1rem;
   color: rgba(216, 216, 247, 0.884);
   text-shadow: 9px -6px 7px rgba(144, 179, 202, 0.64);
 }
