+++ src/index.scss
@@ -14,5 +14,5 @@ body {
   height: 100vh;
   display: flex;
   flex-flow: column nowrap;
-  overflow-y: hidden;
+  overflow: hidden;
 }
