+++ webpack.config.js
@@ -10,7 +10,7 @@ module.exports = {
     new HtmlWebpackPlugin({
       title: 'Pokedex',
       meta: {
-        viewport: 'width=device-width, initial-scale=1, shrink-to-fit=no'
+        viewport: 'width=device-width, initial-scale=1, shrink-to-fit=yes'
       }
     })
   ],
